BIG push for after thanksgiving

+++ HorizonAlpha1/Assets/Scripts/Combat/Damage.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -8,7 +9,9 @@ using UnityEngine;
   {
       [SerializeField] private Collider playerCollider;
 
-      private int damage;
+
+      //private float damageMultiplier = GetComponent<BaseStats>().GetStat(Stat.Damage);
+      private float damage;
       private float knockback;
       public SpriteRenderer sprite;
